+++ extras/UpdatePhenotypes.R
@@ -241,6 +241,12 @@ for (i in (1:length(cohortJsonFiles))) {
     SqlRender::readSql(sourceFile = file.path("inst", "cohorts", jsonFileName))
   sql <-
     CirceR::buildCohortQuery(expression = json, options = circeOptions)
+  writeLines(paste0(" --", sqlFileName))
+  unlink(
+    x = file.path("inst", "sql", "sql_server", sqlFileName),
+    recursive = TRUE,
+    force = TRUE
+  )
   SqlRender::writeSql(
     sql = sql,
     targetFile = file.path("inst", "sql", "sql_server", sqlFileName)
@@ -366,6 +372,59 @@ for (i in (1:nrow(cohortRecord))) {
 
 cohortRecordAugmented <- dplyr::bind_rows(cohortRecordAugmented)
 
+## correct the url
+correctUrl <- function(url) {
+  # Check if the string likely represents a URL by looking for "http"
+  if (grepl("http", url, ignore.case = TRUE)) {
+    # Ensure the URL starts with "https://"
+    if (startsWith(tolower(url), "https//")) {
+      corrected_url <- sub("https//", "https://", url, ignore.case = TRUE)
+      return(corrected_url)
+    }
+    return(url)
+  }
+  return(url)
+}
+
+# Function to transform the URL
+transformUrl <- function(url) {
+  # Check if URL exists
+  if (is.na(url) || url == "" || is.na(url)) {
+    return(url)
+  }
+
+  # Check if URL has the correct base
+  base_url <- "https://forums.ohdsi.org/t/"
+  if (!stringr::str_starts(url, base_url)) {
+    return(NA)
+  }
+
+  extract_number1 <- function(url) {
+    # Use a regular expression to capture the first set of numbers after '/t/' and another '/'
+    match_data <- stringr::str_match(url, "/t/[^/]*/(\\d+)")
+    if (!is.na(match_data[1, 2])) {
+      return(match_data[1, 2])
+    }
+    return(NA)
+  }
+
+  # Extract the first set of numbers after the base URL
+  number1 <- extract_number1(url)
+
+  # If number1 is found, construct the new URL, else return NA
+  if (!is.na(number1)) {
+    new_url <- paste0(base_url, number1)
+    return(new_url)
+  } else {
+    return(url)
+  }
+}
+
+cohortRecordAugmented <- cohortRecordAugmented |>
+  dplyr::mutate(ohdsiForumPost = sapply(ohdsiForumPost, FUN = correctUrl)) |>
+  dplyr::mutate(ohdsiForumPost = sapply(ohdsiForumPost, FUN = transformUrl)) |>
+  dplyr::arrange(cohortId)
+
 readr::write_excel_csv(
   x = cohortRecordAugmented,
   file = "inst/Cohorts.csv",
@@ -378,6 +437,33 @@ if (file.exists("cohortRecord.rds")) {
   file.remove("cohortRecord.rds")
 }
 
+
+newCohortDefinitionSet <-
+  CohortGenerator::getCohortDefinitionSet(
+    settingsFileName = file.path("inst", "Cohorts.csv"),
+    jsonFolder = file.path("inst", "cohorts"),
+    sqlFolder = file.path("inst", "sql", "sql_server")
+  ) |>
+  dplyr::select(
+    cohortId,
+    json
+  ) |>
+  dplyr::tibble() |>
+  dplyr::arrange(cohortId)
+
+conceptSetsInAllCohortDefinition <- ConceptSetDiagnostics::extractConceptSetsInCohortDefinitionSet(
+  cohortDefinitionSet = newCohortDefinitionSet
+)
+
+saveRDS(
+  object = conceptSetsInAllCohortDefinition |>
+    dplyr::arrange(
+      uniqueConceptSetId,
+      cohortId,
+      conceptSetId
+    ),
+  file = file.path("inst", "ConceptSetsInCohortDefinition.RDS")
+)
 
 oldLogFile <- PhenotypeLibrary::getPhenotypeLog()